comfort: 3 clarity: 3
I used mostly my own resources as well as sass-lang.com. I either did well on this one or I am way off, but my understanding is that this stuff works pretty similarly to css so I just applied most of the same understandings. I felt it was straight-forward or very tricky and I am not sure yet...